requirements.txt
Django==4.2.7
djangorestframework==3.14.0
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.7.0